% Generated by roxygen2 (4.0.1): do not edit by hand
\name{answer_list_to_data_frame}
\alias{answer_list_to_data_frame}
\title{convert a list like that assigned from submit_answer to a single row of a data frame}
\usage{
answer_list_to_data_frame(x)
}
\description{
There is some nonstandard stuff here because the Q_value can be
a vector, so I make it a list and use I().
}

