% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkout_hw_branch}
\alias{checkout_hw_branch}
\title{checkout a branch named rrhw-tmp from info in a row of a data frame about open pull requests}
\usage{
checkout_hw_branch(commit)
}
\arguments{
\item{OPR}{A row of a data frame of open pull requests like that
returned by \code{\link{grab_open_pull_requests}}}
}
\description{
This function checks to see if rrhw-tmp is already a branch.  If it is, it will delete the
branch first, and then it will check it out with a new commit. rrhw-tmp should never
get left in a dirty state, or this will bomb.
}

