% Generated by roxygen2 (4.0.1): do not edit by hand
\name{git_pulls_show_to_df}
\alias{git_pulls_show_to_df}
\title{Parse the "git-pulls show" output and return a row of a data frame}
\usage{
git_pulls_show_to_df(N)
}
\arguments{
\item{N}{is the pull request number.}
}
\description{
This also checks to see which files have changed from the merge-base
}

