% Generated by roxygen2 (4.0.1): do not edit by hand
\name{grab_open_pull_requests}
\alias{grab_open_pull_requests}
\title{get all the open pull requests for a branch matching a given pattern}
\usage{
grab_open_pull_requests(branch_p = character(0))
}
\description{
get all the open pull requests for a branch matching a given pattern
}
\examples{
grab_open_pull_requests("^ex-test$")
}

