% Generated by roxygen2 (4.0.1): do not edit by hand
\name{init_homework}
\alias{init_homework}
\title{set variables and knitr hooks for dealing with homeworks}
\usage{
init_homework(homework_name = "Unset")
}
\arguments{
\item{homework_name}{The name you are giving to the homework set}
}
\description{
This should be called in an introductory R block something like this:
which currently is not shown well.  Have to figure out
how to do displayed code with roxygen.
}
\details{
\code{```{r setup, echo=FALSE, include=FALSE}}

\code{rrhw::init_homework("Homework Set 3")}

\code{```}

in an R code block at the
top of the Rmd file you are using this in.
}

