% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pull_requests}
\alias{pull_requests}
\title{return a data frame of pull requests on the repository}
\usage{
pull_requests(branch_p = character(0))
}
\arguments{
\item{state}{Return open, closed, or all pull requests?}

\item{branch_p}{pattern that the branch names should match. This
will filter all pull requests to belong to a certain branch name}
}
\description{
This uses git-pulls
}

