% Generated by roxygen2 (4.0.1): do not edit by hand
\name{submit_answer}
\alias{submit_answer}
\title{workhorse function that processes homeworks}
\usage{
submit_answer(x)
}
\description{
This function is something that the students submit their
answers in, i.e. they write their answers as a (possibly compound)
expression which is an argument to this function. Note that you
must put the curly braces around everthing.
}

